Revert "Add language switcher and locale-prefixed routing (#4)"
This reverts commit d81757e52d7dfffd2498d906565f12e426bec650.
