--- src/app/[lang]/layout.tsx
@@ -1,11 +1,0 @@
-import type { ReactNode } from "react";
-
-import { SUPPORTED_LANGS } from "@/utils/i18n";
-
-export async function generateStaticParams() {
-  return SUPPORTED_LANGS.map((lang) => ({ lang }));
-}
-
-export default function LangLayout({ children }: { children: ReactNode }) {
-  return <>{children}</>;
-}
